Formatting, re-deploys on staging.

+++ contracts/Migrations.sol
@@ -3,21 +3,18 @@ pragma solidity >=0.4.21 <0.6.0;
 
 contract Migrations {
   address public owner;
-  uint public lastCompletedMigration;
+  uint256 public lastCompletedMigration;
 
   constructor() public {
     owner = msg.sender;
   }
 
   modifier restricted() {
-    require(
-      msg.sender == owner,
-      "Only the owner of this contract shall call this method"
-    );
+    require(msg.sender == owner, 'Only the owner of this contract shall call this method');
     _;
   }
 
-  function setCompleted(uint completed) external restricted {
+  function setCompleted(uint256 completed) external restricted {
     lastCompletedMigration = completed;
   }
